+++ BusinessAccountantService/Models/Item.cs
@@ -11,8 +11,8 @@ namespace BusinessAccountantService.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public int Quantity { get; set; }
-        public double PurchasePrice { get; set; }
-        public double RetailPrice { get; set; }
+        public decimal PurchasePrice { get; set; }
+        public decimal RetailPrice { get; set; }
         public string Category {  get; set; }
     }
 }
